added CMakeLists.txt and fixed compilation under windows/without MFC
