+++ cola/libcola/cola_log.h
@@ -161,43 +161,7 @@ class FILELOG_DECLSPEC FILELog : public Log<Output2FILE> {};
     else if (level > cola::FILELog::ReportingLevel() || !cola::Output2FILE::Stream()) ; \
     else cola::FILELog().Get(level)
 
-#if defined(WIN32) || defined(_WIN32) || defined(__WIN32__)
-
-#include <windows.h>
-
-inline std::string NowTime()
-{
-    const int MAX_LEN = 200;
-    char buffer[MAX_LEN];
-    if (GetTimeFormatA(LOCALE_USER_DEFAULT, 0, 0,
-            "HH':'mm':'ss", buffer, MAX_LEN) == 0)
-        return "Error in NowTime()";
-
-    static DWORD first = GetTickCount();
-    std::stringstream result;
-    result << buffer << "." << std::setfill('0') << std::setw(3) << ((long)(GetTickCount() - first) % 1000);
-    return result.str();
-}
-
-#else
-
-#include <sys/time.h>
-
-inline std::string NowTime()
-{
-    char buffer[11];
-    time_t t;
-    time(&t);
-    tm r;
-    strftime(buffer, sizeof(buffer), "%X", localtime_r(&t, &r));
-    struct timeval tv;
-    gettimeofday(&tv, 0);
-    std::stringstream result;
-    result << buffer << "." << std::setfill('0') << std::setw(3) << ((long)tv.tv_usec / 1000);
-    return result.str();
-}
-
-#endif //WIN32
+inline std::string NowTime();
 
 } // namespace cola
