+++ cola/libproject/project.cpp
@@ -22,6 +22,7 @@
 */
 
 #include <algorithm>
+#include <functional>
 #include <numeric>
 #include <cmath>
 #include <cfloat>
